+++ app/view/layout/default.php
@@ -1,7 +1,7 @@
 <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
 <html xmlns="http://www.w3.org/1999/xhtml">
 	<head>
-		<title><?= @$pageTitle ?></title>
+		<title><?php echo @$pageTitle ?></title>
 		<!--[if IE 8]><meta http-equiv="X-UA-Compatible" content="IE=7" /><![endif]-->
 		<?php
 		if (isset($MetaTags)) echo String::indent($MetaTags->render(), 2, TAB, 1);
@@ -10,16 +10,16 @@
 			echo String::indent($CSS->render(), 2, TAB, 1);
 		}
         ?>
-		<link rel="shortcut icon" type="image/ico" href="<?= WEBROOT ?>favicon.ico" />
+		<link rel="shortcut icon" type="image/ico" href="<?php echo WEBROOT ?>favicon.ico" />
 	</head>
 	<body>
 		<div id="app">
 			<div id="header">
-				<h1><a href="<?= WEBROOT ?>">ephFrame <?= ephFrame::VERSION ?></a></h1>
+				<h1><a href="<?php echo WEBROOT ?>">ephFrame <?php echo ephFrame::VERSION ?></a></h1>
 			</div>
-			<?= $this->element('flashMessage') ?>
+			<?php echo $this->element('flashMessage') ?>
 			<div id="content">
-				<?= @$content ?>
+				<?php echo @$content ?>
 			</div>
 			<div id="footer">
 				© 2008+ <a href="http://www.ephigenia.de/" title="Marcel Eichner // Ephigenia" rel="external">Marcel Eichner // Ephigenia</a>,